+++ src/no/hig/imt3591/id3/ID3TestClass.java
@@ -5,32 +5,32 @@ import java.util.LinkedList;
 import java.util.List;
 
 public class ID3TestClass {
-    private static List<Integer[]> observations = new LinkedList<>();
+    private static List<Double[]> observations = new LinkedList<>();
 
     public static void main(String[] args) {
-        observations.add(new Integer[]{0, 0, 0, 10, 0}); //D1
-        observations.add(new Integer[]{0, 0, 1, 7, 0});  //D2
-        observations.add(new Integer[]{1, 0, 0, 6, 1});  //D3
-        observations.add(new Integer[]{2, 0, 0, 4, 1});  //D4
-        observations.add(new Integer[]{2, 1, 0, 9, 1});  //D5
-        observations.add(new Integer[]{2, 1, 1, 8, 0});  //D6
-        observations.add(new Integer[]{1, 1, 1, 2, 1});  //D7
-        observations.add(new Integer[]{0, 0, 0, 10, 0}); //D8
-        observations.add(new Integer[]{0, 1, 0, 5, 1});  //D9
-        observations.add(new Integer[]{2, 1, 0, 5, 1});  //D10
-        observations.add(new Integer[]{0, 1, 1, 5, 1});  //D11
-        observations.add(new Integer[]{1, 0, 1, 5, 1});  //D12
-        observations.add(new Integer[]{1, 1, 0, 5, 1});  //D13
-        observations.add(new Integer[]{2, 0, 1, 10, 0}); //D14
+        observations.add(new Double[]{0d, 0d, 0d, 10d, 0d}); //D1
+        observations.add(new Double[]{0d, 0d, 1d, 7d, 0d});  //D2
+        observations.add(new Double[]{1d, 0d, 0d, 6d, 1d});  //D3
+        observations.add(new Double[]{2d, 0d, 0d, 4d, 1d});  //D4
+        observations.add(new Double[]{2d, 1d, 0d, 9d, 1d});  //D5
+        observations.add(new Double[]{2d, 1d, 1d, 8d, 0d});  //D6
+        observations.add(new Double[]{1d, 1d, 1d, 2d, 1d});  //D7
+        observations.add(new Double[]{0d, 0d, 0d, 10d, 0d}); //D8
+        observations.add(new Double[]{0d, 1d, 0d, 5d, 1d});  //D9
+        observations.add(new Double[]{2d, 1d, 0d, 5d, 1d});  //D10
+        observations.add(new Double[]{0d, 1d, 1d, 5d, 1d});  //D11
+        observations.add(new Double[]{1d, 0d, 1d, 5d, 1d});  //D12
+        observations.add(new Double[]{1d, 1d, 0d, 5d, 1d});  //D13
+        observations.add(new Double[]{2d, 0d, 1d, 10d, 0d}); //D14
 
         IAttributeNameProvider nameProvider = i -> {
                 switch (i) {
                     case 0:
-                        return new Integer[]{0,1,2};
+                        return new Double[]{0d, 1d, 2d};
                     case 1:
-                        return new Integer[]{0,1};
+                        return new Double[]{0d, 1d};
                     case 2:
-                        return new Integer[]{0,1};
+                        return new Double[]{0d, 1d};
                     default:
                         return null;
                 }
@@ -51,8 +51,8 @@ public class ID3TestClass {
         DecisionTree tree = new DecisionTree(observations, 4, 1, nameProvider, metadataProvider);
 
         boolean failed = false;
-        for (Integer[] set : observations) {
-            boolean result = tree.search(new int[]{set[0], set[1], set[2], set[4]});
+        for (Double[] set : observations) {
+            boolean result = tree.search(new double[]{set[0], set[1], set[2], set[4]});
             boolean outputResult = (set[3] == 1);
             if (result != outputResult) {
                 failed = true;
