+++ src/no/hig/imt3591/id3/nodes/AttributeNode.java
@@ -4,23 +4,23 @@ package no.hig.imt3591.id3.nodes;
  * Concrete test on an categorical attributes.
  */
 public class AttributeNode {
-    private int value;
+    private double value;
     private DecisionNode next;
 
-    public AttributeNode(int value, DecisionNode next) {
+    public AttributeNode(double value, DecisionNode next) {
         this.value = value;
         this.next = next;
     }
 
-    public boolean canVisit(int visitor) {
+    public boolean canVisit(double visitor) {
         return visitor == value;
     }
 
-    public boolean visit(int[] set) {
+    public boolean visit(double[] set) {
         return next.search(set);
     }
 
-    protected int getValue() {
+    protected double getValue() {
         return value;
     }
 }
