+++ src/no/hig/imt3591/id3/nodes/BetaLeafNode.java
@@ -4,12 +4,12 @@ package no.hig.imt3591.id3.nodes;
  * A terminal node that goes to the right when it is a terminal/leaf node.
  */
 public class BetaLeafNode extends LeafNode {
-    public BetaLeafNode(int value, boolean endValue) {
+    public BetaLeafNode(double value, boolean endValue) {
         super(value, endValue);
     }
 
     @Override
-    public boolean canVisit(int visitor) {
+    public boolean canVisit(double visitor) {
         return visitor > super.getValue();
     }
 }
