+++ src/no/hig/imt3591/id3/nodes/LeafNode.java
@@ -6,13 +6,13 @@ package no.hig.imt3591.id3.nodes;
 public class LeafNode extends AttributeNode {
     private boolean endValue;
 
-    public LeafNode(int value, boolean endValue) {
+    public LeafNode(double value, boolean endValue) {
         super(value, null);
         this.endValue = endValue;
     }
 
     @Override
-    public boolean visit(int[] set) {
+    public boolean visit(double[] set) {
         return endValue;
     }
 }
